order types with corresponding actions, cleanup
