+++ src/am/components/Orders/Status.js
@@ -15,6 +15,9 @@ const useStyles = makeStyles((theme) => ({
   statusRequested: {
     background: theme.palette.statusRequested.main,
   },
+  statusRejected: {
+    background: theme.palette.statusRejected.main,
+  },
   statusAccepted: {
     background: theme.palette.statusAcceptedTested.main,
   },
@@ -32,6 +35,10 @@ const OrderStatus = ({ orderStatus }) => {
     case 'SubmittedOrder':
       statusText = 'Requested'
       statusClass = `${classes.status} ${classes.statusRequested}`
+      break
+    case 'RejectedOrder':
+      statusText = 'Rejected'
+      statusClass = `${classes.status} ${classes.statusRejected}`
       break
     case 'AcceptedOrder':
       statusText = 'Accepted'
